fixed final pieces from #36 (ivan's PR)
