+++ jax_unirep/utils.py
@@ -289,6 +289,16 @@ def batch_sequences(seqs: List[str]) -> List[List]:
     return order
 
 
+def get_batch_len(batched_seqs: Iterable[str]) -> Tuple[np.ndarray, List]:
+    """
+    Returns the average length of each batch as well as a full array of batch distribution.
+
+    :param batched_seqs: list of lists of sequences, grouped by length.
+    """
+    batch_lens = np.array([len(batch) for batch in batched_seqs])
+    return np.mean(batch_lens), batch_lens
+
+
 # This block of code generates one-hot-encoded arrays.
 oh_arrs = np.eye(max(aa_to_int.values()))
